stdlib: Update gcp.#GCloud to set compute/region instead of compute/zone
If a users correctly sets `region` to a valid GCP region, a warning would be
presented (but is hidden by `-q`):

```
) gcloud config set compute/zone asia-northeast1
Updated property [compute/zone].
WARNING: asia-northeast1 is not a valid zone. Run `gcloud compute zones list` to get all zones.
```

This makes it consistent with AWS where a region is set. However, an optional
`zone` field could also be added to `gcp.#Config`, which would be used to
(conditionally?) set the zone.

AFAICT, there is no option to set a zone in AWS CLI.

Signed-off-by: Scott <slewiskelly@slewiskel.ly>